app/requirements.txt
seqlog>=0.3.31
boto3>=1.26.0
python-dotenv
apscheduler
croniter
pytz>=2023.3